@@ -11,7 +11,8 @@
     "react-dev": "webpack -d --watch",
     "server-dev": "nodemon server/index.js",
     "seed": "node db/createDB.js",
-    "test:client": "jest --watchAll --expand --verbose"
+    "test:client": "jest --watchAll --expand --verbose",
+    "test:coverage": "jest --verbose --coverage"
   },
   "jest": {
     "roots": [
@@ -30,28 +31,39 @@
     ]
   },
   "devDependencies": {
-    "babel-core": "^6.23.1",
+    "@babel/core": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "^23.6.0",
-    "babel-loader": "^6.3.2",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-jest": "^23.2.0",
-    "babel-preset-react": "^6.23.0",
+    "babel-loader": "^8.0.0",
+    "enzyme-adapter-react-16": "^1.1.1",
+    "html-webpack-plugin": "^3.2.0",
     "jest": "^23.6.0",
     "jest-cli": "*",
-    "webpack": "^2.2.1"
+    "jest-dom": "^2.1.0",
+    "react-dom": "^16.4.1",
+    "regenerator-runtime": "^0.12.0",
+    "webpack": "^4.16.1",
+    "webpack-cli": "^3.1.2"
   },
   "dependencies": {
+    "axios": "^0.18.0",
     "body-parser": "^1.17.0",
+    "enzyme": "^3.3.0",
     "express": "^4.15.0",
     "faker": "^4.1.0",
     "file-system": "^2.2.2",
     "jquery": "^3.1.1",
     "mongoose": "^4.13.17",
     "react": "^16.6.0",
-    "react-test-renderer": "^16.6.1",
     "react-bootstrap": "^0.32.4",
     "react-dom": "^16.6.0",
+    "react-test-renderer": "^16.6.1",
+    "react-testing-library": "^5.2.3",
+    "request": "^2.81.0",
     "semantic-ui-react": "^0.83.0",
-    "request": "^2.81.0"
+    "ramda": "^0.25.0",
+    "sinon": "^6.1.3"
   }
 }
