@@ -1,2 +1,2 @@
-var courseData = [{"title":"Multi-layered zero administration initiative","sectionNumber":1,"entries":[{"name":"Talk by Marlen Daniel","duration":196,"entryNumber":1},{"name":"Talk by Quentin Littel","duration":174,"entryNumber":2},{"name":"Talk by Dashawn Tromp","duration":157,"entryNumber":3},{"name":"Talk by Leola Stokes","duration":5,"entryNumber":4},{"name":"Talk by Nola Carter MD","duration":69,"entryNumber":5}]},{"title":"Stand-alone neutral success","sectionNumber":2,"entries":[{"name":"Talk by Orpha Schamberger","duration":252,"entryNumber":1},{"name":"Talk by Maye Nader","duration":228,"entryNumber":2}]},{"title":"Switchable value-added framework","sectionNumber":3,"entries":[{"name":"Talk by Audra Volkman","duration":200,"entryNumber":1},{"name":"Talk by Alia Kuvalis","duration":310,"entryNumber":2},{"name":"Talk by Jesse Skiles","duration":87,"entryNumber":3},{"name":"Talk by Marlee Moen","duration":139,"entryNumber":4},{"name":"Talk by Alfonzo Ernser","duration":42,"entryNumber":5},{"name":"Talk by Johann Vandervort","duration":205,"entryNumber":6},{"name":"Talk by Angelica Dickinson DDS","duration":12,"entryNumber":7},{"name":"Talk by Beatrice Hartmann","duration":70,"entryNumber":8}]},{"title":"Ameliorated static firmware","sectionNumber":4,"entries":[{"name":"Talk by Rachel Howe","duration":76,"entryNumber":1},{"name":"Talk by Erika Wiegand","duration":319,"entryNumber":2},{"name":"Talk by Mr. Vernon Rohan","duration":282,"entryNumber":3},{"name":"Talk by Thurman Von","duration":77,"entryNumber":4},{"name":"Talk by Jaren Sporer","duration":208,"entryNumber":5}]},{"title":"Configurable responsive installation","sectionNumber":5,"entries":[{"name":"Talk by Dora Labadie","duration":91,"entryNumber":1},{"name":"Talk by Carmelo Feil","duration":174,"entryNumber":2},{"name":"Talk by Johanna Prosacco","duration":174,"entryNumber":3},{"name":"Talk by Rhiannon Turcotte","duration":90,"entryNumber":4},{"name":"Talk by Mrs. Everette Grady","duration":335,"entryNumber":5},{"name":"Talk by Theo Renner","duration":174,"entryNumber":6},{"name":"Talk by Donna Moen","duration":335,"entryNumber":7}]},{"title":"Future-proofed 6th generation groupware","sectionNumber":6,"entries":[{"name":"Talk by Aileen Blanda","duration":128,"entryNumber":1},{"name":"Talk by Amya Considine","duration":211,"entryNumber":2},{"name":"Talk by Josie Kozey","duration":88,"entryNumber":3},{"name":"Talk by Serena King","duration":97,"entryNumber":4},{"name":"Talk by Elsie Gislason","duration":139,"entryNumber":5},{"name":"Talk by Malcolm Kautzer","duration":49,"entryNumber":6},{"name":"Talk by Rey Prohaska Sr.","duration":32,"entryNumber":7},{"name":"Talk by Scarlett Nitzsche","duration":326,"entryNumber":8}]},{"title":"Streamlined logistical artificial intelligence","sectionNumber":7,"entries":[{"name":"Talk by Sarina Pfannerstill","duration":65,"entryNumber":1},{"name":"Talk by Maci Langosh","duration":310,"entryNumber":2},{"name":"Talk by Teresa Daniel","duration":116,"entryNumber":3},{"name":"Talk by Allie Feest","duration":356,"entryNumber":4},{"name":"Talk by Lucas Collier","duration":187,"entryNumber":5},{"name":"Talk by Francis Flatley","duration":26,"entryNumber":6},{"name":"Talk by Amir Doyle","duration":58,"entryNumber":7},{"name":"Talk by Mabel Reynolds","duration":319,"entryNumber":8}]},{"title":"Organic disintermediate array","sectionNumber":8,"entries":[{"name":"Talk by Jeffery Jerde","duration":183,"entryNumber":1},{"name":"Talk by Elvis Fadel","duration":204,"entryNumber":2},{"name":"Talk by Valerie Okuneva","duration":36,"entryNumber":3},{"name":"Talk by Marley Yundt","duration":276,"entryNumber":4},{"name":"Talk by Laurence Schaden","duration":111,"entryNumber":5},{"name":"Talk by Malvina Sipes","duration":275,"entryNumber":6},{"name":"Talk by Anastacio Upton","duration":230,"entryNumber":7}]},{"title":"Robust zero administration info-mediaries","sectionNumber":9,"entries":[{"name":"Talk by Shyann Legros","duration":257,"entryNumber":1},{"name":"Talk by Taurean Howell","duration":256,"entryNumber":2},{"name":"Talk by Oma Witting","duration":220,"entryNumber":3},{"name":"Talk by Nayeli Goodwin","duration":124,"entryNumber":4},{"name":"Talk by Dr. Carlie Champlin","duration":101,"entryNumber":5},{"name":"Talk by Yesenia Gutkowski","duration":169,"entryNumber":6},{"name":"Talk by Breanne Zemlak","duration":119,"entryNumber":7}]},{"title":"Visionary directional software","sectionNumber":10,"entries":[{"name":"Talk by Chaya Torphy","duration":33,"entryNumber":1},{"name":"Talk by Vivian Haag","duration":151,"entryNumber":2},{"name":"Talk by Fredy Gusikowski","duration":262,"entryNumber":3},{"name":"Talk by Adolf Kuvalis","duration":48,"entryNumber":4},{"name":"Talk by Margie Leuschke","duration":88,"entryNumber":5},{"name":"Talk by Dr. Amber Hudson","duration":83,"entryNumber":6},{"name":"Talk by Gregoria Schaefer","duration":207,"entryNumber":7}]},{"title":"Universal asymmetric synergy","sectionNumber":11,"entries":[{"name":"Talk by Stacy Mills","duration":225,"entryNumber":1},{"name":"Talk by Ms. Vivienne Gottlieb","duration":162,"entryNumber":2},{"name":"Talk by Dr. Stella Paucek","duration":176,"entryNumber":3}]},{"title":"Multi-lateral multimedia moratorium","sectionNumber":12,"entries":[{"name":"Talk by Hortense Heathcote","duration":291,"entryNumber":1},{"name":"Talk by Gino MacGyver","duration":137,"entryNumber":2},{"name":"Talk by Kailee Yundt","duration":47,"entryNumber":3},{"name":"Talk by Norval O'Kon","duration":336,"entryNumber":4},{"name":"Talk by Adan Gulgowski","duration":52,"entryNumber":5}]},{"title":"Distributed radical intranet","sectionNumber":13,"entries":[{"name":"Talk by Lauretta Bins PhD","duration":328,"entryNumber":1},{"name":"Talk by Mrs. Joe Lehner","duration":40,"entryNumber":2},{"name":"Talk by Lori Rohan I","duration":280,"entryNumber":3},{"name":"Talk by Dayton Botsford I","duration":215,"entryNumber":4},{"name":"Talk by Celine Price","duration":351,"entryNumber":5}]},{"title":"Re-contextualized asynchronous projection","sectionNumber":14,"entries":[{"name":"Talk by Tina Doyle","duration":174,"entryNumber":1},{"name":"Talk by Orie McGlynn","duration":341,"entryNumber":2},{"name":"Talk by Freeman Connelly","duration":101,"entryNumber":3},{"name":"Talk by Nathanial Emard","duration":25,"entryNumber":4}]},{"title":"Programmable fault-tolerant array","sectionNumber":15,"entries":[{"name":"Talk by Miss Dudley Harvey","duration":88,"entryNumber":1},{"name":"Talk by Weldon Roberts","duration":23,"entryNumber":2},{"name":"Talk by Kyla Bergstrom","duration":62,"entryNumber":3},{"name":"Talk by Simone Heidenreich","duration":343,"entryNumber":4},{"name":"Talk by Cindy McClure","duration":170,"entryNumber":5},{"name":"Talk by Raphael Torphy","duration":56,"entryNumber":6},{"name":"Talk by Betty Robel","duration":64,"entryNumber":7}]},{"title":"Re-contextualized radical architecture","sectionNumber":16,"entries":[{"name":"Talk by Henriette Doyle","duration":243,"entryNumber":1},{"name":"Talk by Miss Gregoria Wisozk","duration":109,"entryNumber":2},{"name":"Talk by Darwin Wunsch","duration":91,"entryNumber":3},{"name":"Talk by Jayne Quigley","duration":291,"entryNumber":4},{"name":"Talk by Dr. Katelin Marvin","duration":136,"entryNumber":5},{"name":"Talk by Connor MacGyver","duration":11,"entryNumber":6},{"name":"Talk by Gregory Keeling","duration":40,"entryNumber":7},{"name":"Talk by Roosevelt Connelly","duration":297,"entryNumber":8}]},{"title":"Team-oriented uniform initiative","sectionNumber":17,"entries":[{"name":"Talk by Cristina Howell","duration":12,"entryNumber":1}]},{"title":"Fully-configurable directional monitoring","sectionNumber":18,"entries":[{"name":"Talk by Kaley Aufderhar","duration":232,"entryNumber":1},{"name":"Talk by Donna Adams","duration":200,"entryNumber":2},{"name":"Talk by Ryley Swift","duration":16,"entryNumber":3},{"name":"Talk by Charlene Leffler","duration":174,"entryNumber":4},{"name":"Talk by Earline Larkin","duration":287,"entryNumber":5},{"name":"Talk by Mr. Samir Halvorson","duration":50,"entryNumber":6},{"name":"Talk by Dr. Camden Zieme","duration":265,"entryNumber":7},{"name":"Talk by Heloise McCullough DVM","duration":228,"entryNumber":8}]},{"title":"Expanded demand-driven algorithm","sectionNumber":19,"entries":[{"name":"Talk by Rodrigo Erdman","duration":355,"entryNumber":1},{"name":"Talk by Gerhard Connelly","duration":280,"entryNumber":2},{"name":"Talk by Johnson Muller","duration":23,"entryNumber":3}]},{"title":"Diverse zero tolerance emulation","sectionNumber":20,"entries":[{"name":"Talk by Dr. Margarette Lehner","duration":229,"entryNumber":1},{"name":"Talk by Cara Will","duration":147,"entryNumber":2},{"name":"Talk by Maeve Kshlerin","duration":224,"entryNumber":3},{"name":"Talk by Dawn Denesik","duration":309,"entryNumber":4},{"name":"Talk by Marietta Weber","duration":261,"entryNumber":5},{"name":"Talk by Myriam Morar","duration":346,"entryNumber":6},{"name":"Talk by Henry Schneider","duration":109,"entryNumber":7},{"name":"Talk by Mr. Noe Miller","duration":13,"entryNumber":8}]},{"title":"Enhanced global capacity","sectionNumber":21,"entries":[{"name":"Talk by Juliana Nicolas","duration":287,"entryNumber":1},{"name":"Talk by Bethany Reilly","duration":351,"entryNumber":2},{"name":"Talk by Mrs. Lilyan Volkman","duration":31,"entryNumber":3},{"name":"Talk by Kyle Paucek","duration":279,"entryNumber":4},{"name":"Talk by Mrs. Wilfrid Bailey","duration":275,"entryNumber":5},{"name":"Talk by Joshuah Goyette MD","duration":43,"entryNumber":6},{"name":"Talk by Enrique Koelpin","duration":288,"entryNumber":7},{"name":"Talk by Norval O'Connell","duration":163,"entryNumber":8}]},{"title":"Sharable dynamic knowledge base","sectionNumber":22,"entries":[{"name":"Talk by Jessyca Wiza","duration":71,"entryNumber":1},{"name":"Talk by Della Kihn","duration":93,"entryNumber":2},{"name":"Talk by Mr. Erich Mayert","duration":81,"entryNumber":3}]},{"title":"Versatile clear-thinking Graphical User Interface","sectionNumber":23,"entries":[{"name":"Talk by Herminio Nikolaus V","duration":2,"entryNumber":1},{"name":"Talk by Chad Schaden","duration":281,"entryNumber":2},{"name":"Talk by Madie Stracke","duration":56,"entryNumber":3},{"name":"Talk by Adan Feil","duration":116,"entryNumber":4},{"name":"Talk by Dahlia Bins","duration":342,"entryNumber":5},{"name":"Talk by Lamont Kessler","duration":229,"entryNumber":6},{"name":"Talk by Alia Crooks V","duration":262,"entryNumber":7},{"name":"Talk by Jaquan Dooley","duration":111,"entryNumber":8}]},{"title":"Organic systemic initiative","sectionNumber":24,"entries":[{"name":"Talk by Hilario Rutherford","duration":328,"entryNumber":1},{"name":"Talk by Ines Wunsch","duration":319,"entryNumber":2},{"name":"Talk by Penelope Leuschke","duration":327,"entryNumber":3},{"name":"Talk by Freda Predovic","duration":233,"entryNumber":4}]},{"title":"Managed stable flexibility","sectionNumber":25,"entries":[{"name":"Talk by Damon Bogisich","duration":137,"entryNumber":1},{"name":"Talk by Abel Bednar","duration":336,"entryNumber":2},{"name":"Talk by Jolie Schuppe","duration":347,"entryNumber":3},{"name":"Talk by Emilie Okuneva","duration":172,"entryNumber":4}]},{"title":"Reactive tertiary help-desk","sectionNumber":26,"entries":[{"name":"Talk by Angelina Renner PhD","duration":220,"entryNumber":1},{"name":"Talk by Ms. Keshawn Crooks","duration":96,"entryNumber":2},{"name":"Talk by Sibyl Bernier","duration":340,"entryNumber":3},{"name":"Talk by Derick Daugherty MD","duration":127,"entryNumber":4},{"name":"Talk by Eriberto Christiansen","duration":352,"entryNumber":5},{"name":"Talk by Dr. Walter Yundt","duration":259,"entryNumber":6}]},{"title":"Compatible solution-oriented synergy","sectionNumber":27,"entries":[{"name":"Talk by Matilda Rogahn","duration":287,"entryNumber":1},{"name":"Talk by Ms. Bernardo Kovacek","duration":55,"entryNumber":2},{"name":"Talk by Oswaldo Eichmann","duration":285,"entryNumber":3},{"name":"Talk by Emie Reinger","duration":275,"entryNumber":4},{"name":"Talk by Vaughn Rogahn","duration":332,"entryNumber":5},{"name":"Talk by Alek Wunsch","duration":321,"entryNumber":6}]},{"title":"Optimized bi-directional success","sectionNumber":28,"entries":[{"name":"Talk by Vella Rempel","duration":256,"entryNumber":1},{"name":"Talk by Rafaela Greenholt","duration":115,"entryNumber":2},{"name":"Talk by Trey Jacobs","duration":326,"entryNumber":3},{"name":"Talk by Ms. Lynn Wiza","duration":343,"entryNumber":4},{"name":"Talk by Salvador Lueilwitz","duration":56,"entryNumber":5},{"name":"Talk by Ryan Toy","duration":329,"entryNumber":6}]},{"title":"Horizontal mobile application","sectionNumber":29,"entries":[{"name":"Talk by Lou Jacobi","duration":41,"entryNumber":1},{"name":"Talk by Miss Chloe Cartwright","duration":333,"entryNumber":2},{"name":"Talk by Narciso Lebsack DDS","duration":119,"entryNumber":3},{"name":"Talk by Shayne Kunze I","duration":7,"entryNumber":4},{"name":"Talk by Isidro Senger","duration":144,"entryNumber":5}]},{"title":"Customer-focused disintermediate complexity","sectionNumber":30,"entries":[{"name":"Talk by Carole Tromp","duration":26,"entryNumber":1}]},{"title":"Team-oriented incremental hardware","sectionNumber":31,"entries":[{"name":"Talk by Mikayla Koch","duration":267,"entryNumber":1},{"name":"Talk by Anastasia Quitzon","duration":302,"entryNumber":2},{"name":"Talk by Karelle Renner","duration":54,"entryNumber":3},{"name":"Talk by Daron Nolan","duration":70,"entryNumber":4}]},{"title":"Cloned clear-thinking local area network","sectionNumber":32,"entries":[{"name":"Talk by Allene Batz","duration":231,"entryNumber":1}]},{"title":"Configurable stable pricing structure","sectionNumber":33,"entries":[{"name":"Talk by Arely Fritsch","duration":19,"entryNumber":1},{"name":"Talk by Ethelyn Farrell","duration":116,"entryNumber":2},{"name":"Talk by August Leuschke","duration":339,"entryNumber":3},{"name":"Talk by Mr. Alba Nicolas","duration":236,"entryNumber":4},{"name":"Talk by Miss Marina Johnston","duration":187,"entryNumber":5},{"name":"Talk by Bettye Ankunding","duration":34,"entryNumber":6}]},{"title":"Extended client-driven process improvement","sectionNumber":34,"entries":[{"name":"Talk by Devante Armstrong","duration":229,"entryNumber":1},{"name":"Talk by Dr. Nona Kilback","duration":42,"entryNumber":2},{"name":"Talk by Madaline Rohan","duration":231,"entryNumber":3},{"name":"Talk by Clyde Rath","duration":133,"entryNumber":4},{"name":"Talk by Erling Russel","duration":1,"entryNumber":5}]},{"title":"Future-proofed bandwidth-monitored instruction set","sectionNumber":35,"entries":[{"name":"Talk by Raegan Ward MD","duration":161,"entryNumber":1},{"name":"Talk by Akeem Brown","duration":9,"entryNumber":2},{"name":"Talk by Immanuel Berge","duration":44,"entryNumber":3},{"name":"Talk by Roxane Zieme","duration":157,"entryNumber":4},{"name":"Talk by Emily Koch","duration":159,"entryNumber":5},{"name":"Talk by Mrs. Antonietta Nader","duration":17,"entryNumber":6}]},{"title":"Face to face reciprocal synergy","sectionNumber":36,"entries":[{"name":"Talk by Ms. Thea Corwin","duration":32,"entryNumber":1},{"name":"Talk by Casimir Ratke","duration":238,"entryNumber":2},{"name":"Talk by Durward Littel DDS","duration":208,"entryNumber":3},{"name":"Talk by Opal Hirthe","duration":276,"entryNumber":4},{"name":"Talk by Piper Stanton","duration":186,"entryNumber":5},{"name":"Talk by Desmond Schowalter","duration":274,"entryNumber":6},{"name":"Talk by Jerry Mueller","duration":88,"entryNumber":7},{"name":"Talk by Armani Runte","duration":271,"entryNumber":8}]},{"title":"Universal asymmetric portal","sectionNumber":37,"entries":[{"name":"Talk by Darren Weissnat","duration":52,"entryNumber":1},{"name":"Talk by Liliana Torphy","duration":58,"entryNumber":2},{"name":"Talk by Seamus Carter","duration":41,"entryNumber":3},{"name":"Talk by Tyreek Hermann","duration":147,"entryNumber":4},{"name":"Talk by Antonia Borer","duration":338,"entryNumber":5},{"name":"Talk by Izaiah Schneider","duration":117,"entryNumber":6}]},{"title":"Intuitive 6th generation definition","sectionNumber":38,"entries":[{"name":"Talk by Emma Murphy","duration":240,"entryNumber":1},{"name":"Talk by Evalyn Gislason","duration":42,"entryNumber":2},{"name":"Talk by Verdie Krajcik","duration":234,"entryNumber":3}]},{"title":"Persevering needs-based open architecture","sectionNumber":39,"entries":[{"name":"Talk by Zita Hoppe","duration":271,"entryNumber":1},{"name":"Talk by Mr. Jewel Mitchell","duration":43,"entryNumber":2},{"name":"Talk by Leonel Hartmann","duration":290,"entryNumber":3}]},{"title":"Business-focused bifurcated matrix","sectionNumber":40,"entries":[{"name":"Talk by Nelson Von","duration":279,"entryNumber":1},{"name":"Talk by Enrico D'Amore IV","duration":139,"entryNumber":2}]},{"title":"Mandatory 6th generation synergy","sectionNumber":41,"entries":[{"name":"Talk by Bernhard Mante II","duration":350,"entryNumber":1},{"name":"Talk by Mckayla Zemlak","duration":20,"entryNumber":2},{"name":"Talk by Emelia Hickle","duration":339,"entryNumber":3},{"name":"Talk by Harmon Mohr","duration":323,"entryNumber":4},{"name":"Talk by Miss Lavonne Wilderman","duration":88,"entryNumber":5},{"name":"Talk by Oleta Haley","duration":238,"entryNumber":6},{"name":"Talk by Dr. Donald Monahan","duration":84,"entryNumber":7},{"name":"Talk by Tara O'Hara","duration":99,"entryNumber":8}]},{"title":"Fully-configurable 5th generation concept","sectionNumber":42,"entries":[{"name":"Talk by Jackson Bogan DDS","duration":321,"entryNumber":1},{"name":"Talk by Dannie Jerde DVM","duration":138,"entryNumber":2},{"name":"Talk by Beatrice Herzog","duration":114,"entryNumber":3},{"name":"Talk by Isabella Rowe","duration":307,"entryNumber":4},{"name":"Talk by Mac Skiles","duration":118,"entryNumber":5},{"name":"Talk by Demario Adams","duration":56,"entryNumber":6}]},{"title":"Organic modular parallelism","sectionNumber":43,"entries":[{"name":"Talk by Raul Cummerata","duration":7,"entryNumber":1},{"name":"Talk by Baron Stroman","duration":211,"entryNumber":2},{"name":"Talk by Maria Schultz","duration":358,"entryNumber":3},{"name":"Talk by Mr. Haylie Bruen","duration":56,"entryNumber":4},{"name":"Talk by Gabriel Fritsch","duration":271,"entryNumber":5},{"name":"Talk by Wanda Olson","duration":0,"entryNumber":6},{"name":"Talk by Catharine Turner","duration":88,"entryNumber":7},{"name":"Talk by Mr. Blaze Parisian","duration":62,"entryNumber":8}]},{"title":"De-engineered methodical monitoring","sectionNumber":44,"entries":[{"name":"Talk by Noe Koch","duration":125,"entryNumber":1},{"name":"Talk by Wilmer Keeling","duration":32,"entryNumber":2},{"name":"Talk by Catalina Stokes","duration":286,"entryNumber":3},{"name":"Talk by Sonya Swift","duration":357,"entryNumber":4},{"name":"Talk by Greta Feil","duration":247,"entryNumber":5},{"name":"Talk by Candelario Cummerata MD","duration":205,"entryNumber":6},{"name":"Talk by Tyree Hayes","duration":37,"entryNumber":7}]},{"title":"Persistent foreground internet solution","sectionNumber":45,"entries":[{"name":"Talk by Kareem Hammes","duration":16,"entryNumber":1},{"name":"Talk by Summer Gibson","duration":84,"entryNumber":2},{"name":"Talk by Jacinto Steuber","duration":288,"entryNumber":3},{"name":"Talk by Kyler Strosin DDS","duration":301,"entryNumber":4},{"name":"Talk by Ardith Keeling","duration":331,"entryNumber":5},{"name":"Talk by Derek Schiller","duration":293,"entryNumber":6},{"name":"Talk by Mr. Sarah Pollich","duration":187,"entryNumber":7}]},{"title":"Diverse regional core","sectionNumber":46,"entries":[{"name":"Talk by Kattie Feil","duration":134,"entryNumber":1},{"name":"Talk by Newton Schimmel Jr.","duration":55,"entryNumber":2}]},{"title":"Re-engineered encompassing knowledge base","sectionNumber":47,"entries":[{"name":"Talk by Jarvis Quigley III","duration":278,"entryNumber":1},{"name":"Talk by Linwood Harris","duration":199,"entryNumber":2},{"name":"Talk by Mr. Loren Donnelly","duration":54,"entryNumber":3},{"name":"Talk by Marcelino Beer","duration":264,"entryNumber":4},{"name":"Talk by Stevie Spencer Jr.","duration":46,"entryNumber":5}]},{"title":"Mandatory reciprocal adapter","sectionNumber":48,"entries":[{"name":"Talk by Reid Beier","duration":335,"entryNumber":1},{"name":"Talk by Augustus Dickens PhD","duration":318,"entryNumber":2},{"name":"Talk by Kara Turcotte","duration":77,"entryNumber":3},{"name":"Talk by Hallie Stoltenberg","duration":46,"entryNumber":4},{"name":"Talk by Chyna Robel","duration":144,"entryNumber":5},{"name":"Talk by Dr. Annamae Howe","duration":170,"entryNumber":6},{"name":"Talk by Zoie Prohaska","duration":59,"entryNumber":7}]},{"title":"Distributed 3rd generation focus group","sectionNumber":49,"entries":[{"name":"Talk by Martina King","duration":335,"entryNumber":1},{"name":"Talk by Monica Kuvalis","duration":350,"entryNumber":2},{"name":"Talk by Eden Huels","duration":97,"entryNumber":3},{"name":"Talk by Jarod Tromp","duration":51,"entryNumber":4},{"name":"Talk by Angie Harvey","duration":231,"entryNumber":5},{"name":"Talk by Mariela O'Hara","duration":80,"entryNumber":6},{"name":"Talk by Harry Auer","duration":71,"entryNumber":7},{"name":"Talk by Clementina Champlin","duration":320,"entryNumber":8}]},{"title":"Devolved asynchronous neural-net","sectionNumber":50,"entries":[{"name":"Talk by Clementina Gleichner","duration":284,"entryNumber":1},{"name":"Talk by Devyn Zboncak","duration":314,"entryNumber":2},{"name":"Talk by Mr. Bella Collier","duration":60,"entryNumber":3}]},{"title":"Exclusive intermediate initiative","sectionNumber":51,"entries":[{"name":"Talk by Jorge Mayer","duration":57,"entryNumber":1}]},{"title":"Automated coherent methodology","sectionNumber":52,"entries":[{"name":"Talk by Yolanda Muller","duration":349,"entryNumber":1},{"name":"Talk by Alda Bednar","duration":31,"entryNumber":2},{"name":"Talk by Karlee Eichmann","duration":245,"entryNumber":3},{"name":"Talk by Clara Crist","duration":288,"entryNumber":4},{"name":"Talk by Millie Mraz","duration":232,"entryNumber":5},{"name":"Talk by Rogers Grady","duration":60,"entryNumber":6},{"name":"Talk by Emanuel Wintheiser","duration":14,"entryNumber":7}]},{"title":"Extended mobile success","sectionNumber":53,"entries":[{"name":"Talk by Bulah Wolf","duration":280,"entryNumber":1}]},{"title":"Front-line intermediate projection","sectionNumber":54,"entries":[{"name":"Talk by Karen Langosh I","duration":7,"entryNumber":1},{"name":"Talk by Aleen Zieme","duration":118,"entryNumber":2},{"name":"Talk by Rodrick Mohr","duration":340,"entryNumber":3}]},{"title":"Enhanced bifurcated flexibility","sectionNumber":55,"entries":[{"name":"Talk by Betsy Hansen","duration":29,"entryNumber":1},{"name":"Talk by Isaiah Daniel","duration":191,"entryNumber":2},{"name":"Talk by Mertie Deckow","duration":100,"entryNumber":3}]},{"title":"Expanded directional pricing structure","sectionNumber":56,"entries":[{"name":"Talk by Mr. Oda Herzog","duration":136,"entryNumber":1},{"name":"Talk by Haylie Nolan","duration":213,"entryNumber":2}]},{"title":"Vision-oriented encompassing focus group","sectionNumber":57,"entries":[{"name":"Talk by Reese Fay","duration":110,"entryNumber":1},{"name":"Talk by Violette Nikolaus","duration":229,"entryNumber":2},{"name":"Talk by Rogers MacGyver","duration":120,"entryNumber":3},{"name":"Talk by Winnifred Jenkins","duration":129,"entryNumber":4},{"name":"Talk by Kole Ernser","duration":293,"entryNumber":5},{"name":"Talk by Bettie Orn","duration":25,"entryNumber":6},{"name":"Talk by Jerrold Bradtke","duration":333,"entryNumber":7}]},{"title":"Virtual scalable help-desk","sectionNumber":58,"entries":[{"name":"Talk by Nicholaus Lockman","duration":142,"entryNumber":1},{"name":"Talk by Everette Hayes IV","duration":131,"entryNumber":2},{"name":"Talk by Emmanuel O'Conner Jr.","duration":243,"entryNumber":3},{"name":"Talk by Madilyn Beahan","duration":46,"entryNumber":4},{"name":"Talk by Oda Predovic","duration":211,"entryNumber":5},{"name":"Talk by Mr. Mariane Herzog","duration":132,"entryNumber":6},{"name":"Talk by Napoleon Cummerata","duration":181,"entryNumber":7},{"name":"Talk by Kristy Upton","duration":82,"entryNumber":8}]},{"title":"Balanced intangible strategy","sectionNumber":59,"entries":[{"name":"Talk by Russel Turcotte","duration":74,"entryNumber":1},{"name":"Talk by Juliana Beer DVM","duration":322,"entryNumber":2},{"name":"Talk by Lance Rempel","duration":204,"entryNumber":3},{"name":"Talk by Oliver Fay","duration":251,"entryNumber":4}]},{"title":"Face to face contextually-based concept","sectionNumber":60,"entries":[{"name":"Talk by Verla Dooley","duration":80,"entryNumber":1},{"name":"Talk by Joan Hermann","duration":349,"entryNumber":2},{"name":"Talk by Skyla Raynor MD","duration":69,"entryNumber":3}]},{"title":"Persistent non-volatile standardization","sectionNumber":61,"entries":[{"name":"Talk by Annabel Bosco","duration":278,"entryNumber":1},{"name":"Talk by Darrell Koepp","duration":48,"entryNumber":2},{"name":"Talk by Giovanna Miller","duration":62,"entryNumber":3}]},{"title":"Enhanced web-enabled archive","sectionNumber":62,"entries":[{"name":"Talk by Madeline Moore","duration":102,"entryNumber":1},{"name":"Talk by Yesenia Kshlerin","duration":208,"entryNumber":2},{"name":"Talk by Torrey Sipes","duration":77,"entryNumber":3},{"name":"Talk by Destin Johnston","duration":94,"entryNumber":4},{"name":"Talk by Ally Bauch V","duration":173,"entryNumber":5},{"name":"Talk by Lorine Heathcote","duration":92,"entryNumber":6}]},{"title":"Organized secondary focus group","sectionNumber":63,"entries":[{"name":"Talk by Lucious Schowalter","duration":136,"entryNumber":1},{"name":"Talk by Brain Kihn","duration":126,"entryNumber":2},{"name":"Talk by Tevin Haag V","duration":1,"entryNumber":3}]},{"title":"Function-based bottom-line algorithm","sectionNumber":64,"entries":[{"name":"Talk by Shaylee Kuphal","duration":142,"entryNumber":1},{"name":"Talk by Dr. Esperanza Greenfelder","duration":358,"entryNumber":2},{"name":"Talk by Arnaldo Beier DDS","duration":90,"entryNumber":3},{"name":"Talk by Ulises Miller","duration":68,"entryNumber":4}]},{"title":"Networked needs-based intranet","sectionNumber":65,"entries":[{"name":"Talk by Lambert Greenfelder","duration":305,"entryNumber":1}]},{"title":"Re-contextualized zero tolerance algorithm","sectionNumber":66,"entries":[{"name":"Talk by Eva Brekke","duration":112,"entryNumber":1},{"name":"Talk by Delilah Lynch","duration":300,"entryNumber":2},{"name":"Talk by Micaela Nolan","duration":267,"entryNumber":3},{"name":"Talk by Kaia Denesik","duration":315,"entryNumber":4}]},{"title":"Managed 3rd generation concept","sectionNumber":67,"entries":[{"name":"Talk by Leonie Hane","duration":136,"entryNumber":1},{"name":"Talk by Trinity Braun II","duration":160,"entryNumber":2}]},{"title":"Realigned heuristic policy","sectionNumber":68,"entries":[{"name":"Talk by Joanie Sawayn","duration":291,"entryNumber":1},{"name":"Talk by Clementina Steuber","duration":94,"entryNumber":2},{"name":"Talk by Norene Runte","duration":193,"entryNumber":3},{"name":"Talk by Sven Shanahan","duration":136,"entryNumber":4}]},{"title":"Diverse bandwidth-monitored Graphical User Interface","sectionNumber":69,"entries":[{"name":"Talk by Norma Walker","duration":69,"entryNumber":1}]},{"title":"Assimilated multi-tasking encryption","sectionNumber":70,"entries":[{"name":"Talk by Ms. Retha Wolff","duration":334,"entryNumber":1},{"name":"Talk by Jaquelin O'Hara","duration":65,"entryNumber":2},{"name":"Talk by Hannah Dare","duration":140,"entryNumber":3}]},{"title":"Enhanced non-volatile ability","sectionNumber":71,"entries":[{"name":"Talk by Koby Connelly","duration":38,"entryNumber":1}]},{"title":"Self-enabling fault-tolerant moderator","sectionNumber":72,"entries":[{"name":"Talk by Carey Gusikowski","duration":197,"entryNumber":1},{"name":"Talk by Gonzalo Kuvalis","duration":338,"entryNumber":2},{"name":"Talk by Lisette Mante","duration":352,"entryNumber":3},{"name":"Talk by Drew Huel","duration":301,"entryNumber":4}]},{"title":"Integrated systemic adapter","sectionNumber":73,"entries":[{"name":"Talk by Andy Rice","duration":236,"entryNumber":1},{"name":"Talk by Jayce Lemke","duration":33,"entryNumber":2},{"name":"Talk by Ms. Felipe Gibson","duration":134,"entryNumber":3},{"name":"Talk by Guido Mante","duration":305,"entryNumber":4}]},{"title":"Reactive asynchronous structure","sectionNumber":74,"entries":[{"name":"Talk by Merl Quitzon","duration":39,"entryNumber":1},{"name":"Talk by Brielle Sporer Jr.","duration":163,"entryNumber":2},{"name":"Talk by Bertrand Hoppe","duration":0,"entryNumber":3},{"name":"Talk by Raina Hermiston DDS","duration":187,"entryNumber":4}]},{"title":"Polarised bottom-line frame","sectionNumber":75,"entries":[{"name":"Talk by Wilma Yundt","duration":213,"entryNumber":1},{"name":"Talk by Candido Quitzon","duration":205,"entryNumber":2},{"name":"Talk by Taya Price","duration":25,"entryNumber":3}]},{"title":"Synchronised holistic open system","sectionNumber":76,"entries":[{"name":"Talk by Joaquin Gerhold","duration":289,"entryNumber":1},{"name":"Talk by Mr. Mazie Torp","duration":164,"entryNumber":2},{"name":"Talk by Tatyana Reilly","duration":238,"entryNumber":3},{"name":"Talk by Hallie Bauch","duration":269,"entryNumber":4},{"name":"Talk by Bridie O'Reilly","duration":151,"entryNumber":5},{"name":"Talk by Shanel Harber","duration":33,"entryNumber":6}]},{"title":"Profound asynchronous neural-net","sectionNumber":77,"entries":[{"name":"Talk by Alia Strosin","duration":59,"entryNumber":1},{"name":"Talk by Minerva Nader DVM","duration":13,"entryNumber":2},{"name":"Talk by Jaden Fadel Jr.","duration":102,"entryNumber":3}]},{"title":"Optional responsive archive","sectionNumber":78,"entries":[{"name":"Talk by Grady Kassulke","duration":272,"entryNumber":1},{"name":"Talk by Wilhelm Buckridge","duration":350,"entryNumber":2}]},{"title":"Self-enabling 3rd generation system engine","sectionNumber":79,"entries":[{"name":"Talk by Terence Hamill","duration":182,"entryNumber":1},{"name":"Talk by Norval Botsford","duration":17,"entryNumber":2}]},{"title":"Integrated bifurcated project","sectionNumber":80,"entries":[{"name":"Talk by Ms. Elisa Fahey","duration":85,"entryNumber":1},{"name":"Talk by Mackenzie Cassin","duration":13,"entryNumber":2},{"name":"Talk by Aisha King","duration":215,"entryNumber":3},{"name":"Talk by Miss Mario Wolf","duration":81,"entryNumber":4}]},{"title":"Total zero defect function","sectionNumber":81,"entries":[{"name":"Talk by Miss Korey Prosacco","duration":166,"entryNumber":1},{"name":"Talk by Ollie Pacocha","duration":310,"entryNumber":2},{"name":"Talk by Madonna Nicolas","duration":311,"entryNumber":3},{"name":"Talk by Oda Hickle","duration":141,"entryNumber":4},{"name":"Talk by Chauncey Doyle","duration":275,"entryNumber":5},{"name":"Talk by Miss Agustin Blick","duration":45,"entryNumber":6},{"name":"Talk by Demario Jaskolski","duration":354,"entryNumber":7},{"name":"Talk by Justyn D'Amore","duration":281,"entryNumber":8}]},{"title":"Realigned foreground utilisation","sectionNumber":82,"entries":[{"name":"Talk by Salvador Heller","duration":266,"entryNumber":1},{"name":"Talk by Nikki Boyer","duration":26,"entryNumber":2},{"name":"Talk by Josefa Prosacco","duration":91,"entryNumber":3},{"name":"Talk by Cordia Feil","duration":15,"entryNumber":4},{"name":"Talk by Margaret Schiller","duration":310,"entryNumber":5}]},{"title":"Integrated maximized definition","sectionNumber":83,"entries":[{"name":"Talk by Toby Hartmann","duration":209,"entryNumber":1},{"name":"Talk by Maryjane Nitzsche","duration":139,"entryNumber":2}]},{"title":"Stand-alone dedicated contingency","sectionNumber":84,"entries":[{"name":"Talk by Shanny Nienow","duration":340,"entryNumber":1},{"name":"Talk by Hellen Howe","duration":250,"entryNumber":2},{"name":"Talk by Ali Jerde","duration":302,"entryNumber":3},{"name":"Talk by Effie Zemlak","duration":38,"entryNumber":4},{"name":"Talk by Reynold Hodkiewicz","duration":311,"entryNumber":5},{"name":"Talk by Mr. Clint Dickens","duration":95,"entryNumber":6},{"name":"Talk by Quinten Nitzsche Jr.","duration":90,"entryNumber":7}]},{"title":"Sharable mission-critical extranet","sectionNumber":85,"entries":[{"name":"Talk by Brook Schaefer","duration":62,"entryNumber":1},{"name":"Talk by Kolby Gutkowski IV","duration":295,"entryNumber":2},{"name":"Talk by Ransom Cartwright","duration":134,"entryNumber":3},{"name":"Talk by Sydni Blick","duration":349,"entryNumber":4},{"name":"Talk by Warren Bosco","duration":54,"entryNumber":5},{"name":"Talk by Raymundo Conn","duration":212,"entryNumber":6},{"name":"Talk by Lilly Crona I","duration":320,"entryNumber":7},{"name":"Talk by Cornelius Kris","duration":60,"entryNumber":8}]},{"title":"Re-engineered homogeneous collaboration","sectionNumber":86,"entries":[{"name":"Talk by Walton Dach","duration":14,"entryNumber":1},{"name":"Talk by Mrs. Mustafa Crooks","duration":156,"entryNumber":2},{"name":"Talk by Carmel Bogan","duration":154,"entryNumber":3},{"name":"Talk by Eduardo Bogan","duration":147,"entryNumber":4},{"name":"Talk by Izaiah Breitenberg","duration":21,"entryNumber":5},{"name":"Talk by Cecilia Ankunding","duration":330,"entryNumber":6},{"name":"Talk by Charity McGlynn","duration":131,"entryNumber":7},{"name":"Talk by Alexis Crona","duration":43,"entryNumber":8}]},{"title":"Multi-channelled hybrid frame","sectionNumber":87,"entries":[{"name":"Talk by Sonya Herzog","duration":163,"entryNumber":1},{"name":"Talk by Rhea Quitzon","duration":113,"entryNumber":2},{"name":"Talk by Jordan Terry","duration":85,"entryNumber":3},{"name":"Talk by Hobart Beer","duration":269,"entryNumber":4},{"name":"Talk by Mr. Amya Welch","duration":318,"entryNumber":5},{"name":"Talk by Virgie Moore","duration":133,"entryNumber":6},{"name":"Talk by Aiden Crist","duration":143,"entryNumber":7}]},{"title":"Advanced exuding help-desk","sectionNumber":88,"entries":[{"name":"Talk by Dashawn Hickle","duration":120,"entryNumber":1},{"name":"Talk by Ms. Michele Grant","duration":192,"entryNumber":2},{"name":"Talk by Adeline Blanda","duration":217,"entryNumber":3},{"name":"Talk by Ryleigh Haag","duration":131,"entryNumber":4},{"name":"Talk by Maureen Conroy","duration":127,"entryNumber":5},{"name":"Talk by Morris Hand","duration":87,"entryNumber":6},{"name":"Talk by Zoie Jenkins","duration":330,"entryNumber":7},{"name":"Talk by Pete Daniel","duration":127,"entryNumber":8}]},{"title":"Front-line intangible capability","sectionNumber":89,"entries":[{"name":"Talk by Jakayla Pfannerstill","duration":77,"entryNumber":1},{"name":"Talk by Eladio Schneider","duration":158,"entryNumber":2},{"name":"Talk by Pascale Kunze","duration":187,"entryNumber":3}]},{"title":"Persistent tertiary open architecture","sectionNumber":90,"entries":[{"name":"Talk by Geovanny Auer","duration":235,"entryNumber":1},{"name":"Talk by Zoila Swaniawski","duration":290,"entryNumber":2},{"name":"Talk by Jacquelyn Johns","duration":219,"entryNumber":3}]},{"title":"Implemented secondary policy","sectionNumber":91,"entries":[{"name":"Talk by Kathlyn Stark","duration":202,"entryNumber":1},{"name":"Talk by Janice Gerlach","duration":243,"entryNumber":2}]},{"title":"Multi-lateral dynamic projection","sectionNumber":92,"entries":[{"name":"Talk by Lorena Rempel","duration":32,"entryNumber":1},{"name":"Talk by Peyton Connelly","duration":139,"entryNumber":2},{"name":"Talk by Waylon Corkery","duration":115,"entryNumber":3},{"name":"Talk by Dora Stroman PhD","duration":276,"entryNumber":4},{"name":"Talk by Herta Runolfsson","duration":195,"entryNumber":5},{"name":"Talk by Eileen Flatley","duration":175,"entryNumber":6},{"name":"Talk by Eli Reichert","duration":353,"entryNumber":7}]},{"title":"Virtual dynamic leverage","sectionNumber":93,"entries":[{"name":"Talk by Ramiro Ortiz","duration":171,"entryNumber":1},{"name":"Talk by Sigrid Hilll","duration":206,"entryNumber":2},{"name":"Talk by Gregoria Beier","duration":57,"entryNumber":3},{"name":"Talk by Miss Aubrey Littel","duration":235,"entryNumber":4},{"name":"Talk by Oleta Ritchie","duration":349,"entryNumber":5},{"name":"Talk by Demetris Wiegand IV","duration":354,"entryNumber":6}]},{"title":"Proactive well-modulated conglomeration","sectionNumber":94,"entries":[{"name":"Talk by Gideon Beer","duration":249,"entryNumber":1}]},{"title":"Diverse user-facing forecast","sectionNumber":95,"entries":[{"name":"Talk by Georgiana Abbott","duration":239,"entryNumber":1}]},{"title":"Enterprise-wide methodical encryption","sectionNumber":96,"entries":[{"name":"Talk by Mercedes Grimes I","duration":65,"entryNumber":1},{"name":"Talk by Lysanne Hahn","duration":225,"entryNumber":2},{"name":"Talk by Lizzie Kling","duration":256,"entryNumber":3}]},{"title":"Synchronised bi-directional artificial intelligence","sectionNumber":97,"entries":[{"name":"Talk by Caleb Schmitt","duration":79,"entryNumber":1},{"name":"Talk by Dr. Alexandrea Pouros","duration":140,"entryNumber":2},{"name":"Talk by Marquis Thompson","duration":187,"entryNumber":3},{"name":"Talk by Marques Weimann","duration":203,"entryNumber":4},{"name":"Talk by Elian Ryan","duration":185,"entryNumber":5},{"name":"Talk by Sienna Gerlach","duration":311,"entryNumber":6},{"name":"Talk by Virginie Upton","duration":184,"entryNumber":7}]},{"title":"Future-proofed even-keeled model","sectionNumber":98,"entries":[{"name":"Talk by Jesus Fritsch","duration":142,"entryNumber":1},{"name":"Talk by Beryl Connelly","duration":346,"entryNumber":2},{"name":"Talk by Camden Padberg DVM","duration":347,"entryNumber":3},{"name":"Talk by Shyann Altenwerth","duration":15,"entryNumber":4},{"name":"Talk by Christelle Hauck IV","duration":65,"entryNumber":5},{"name":"Talk by Narciso Cole","duration":258,"entryNumber":6}]},{"title":"Managed national solution","sectionNumber":99,"entries":[{"name":"Talk by Miss Juanita Hills","duration":39,"entryNumber":1},{"name":"Talk by Emerald Kuhic","duration":294,"entryNumber":2},{"name":"Talk by Wilburn Mayert","duration":342,"entryNumber":3},{"name":"Talk by Kelsie Green I","duration":310,"entryNumber":4},{"name":"Talk by Josephine Deckow","duration":304,"entryNumber":5},{"name":"Talk by Weston Schulist","duration":338,"entryNumber":6}]},{"title":"Integrated logistical open architecture","sectionNumber":100,"entries":[{"name":"Talk by Shaina Block","duration":119,"entryNumber":1},{"name":"Talk by Maggie Schowalter","duration":104,"entryNumber":2},{"name":"Talk by Luther Prosacco Sr.","duration":0,"entryNumber":3},{"name":"Talk by Catherine Legros","duration":11,"entryNumber":4}]}] 
+var courseData = [{"title":"Progressive 24/7 Graphical User Interface","sectionNumber":1,"entries":[{"name":"Talk by Agustin Murray","duration":249,"entryNumber":1}]},{"title":"Switchable attitude-oriented throughput","sectionNumber":2,"entries":[{"name":"Talk by Tabitha O'Connell","duration":69,"entryNumber":1},{"name":"Talk by Hertha McDermott","duration":45,"entryNumber":2},{"name":"Talk by Bruce Strosin","duration":45,"entryNumber":3},{"name":"Talk by Dr. Laila Ebert","duration":265,"entryNumber":4},{"name":"Talk by Daisha Hilll Sr.","duration":204,"entryNumber":5}]},{"title":"Stand-alone actuating groupware","sectionNumber":3,"entries":[{"name":"Talk by Mrs. Effie Terry","duration":138,"entryNumber":1},{"name":"Talk by Burley Ward","duration":223,"entryNumber":2},{"name":"Talk by Travis Cartwright","duration":350,"entryNumber":3},{"name":"Talk by Dexter Luettgen","duration":165,"entryNumber":4},{"name":"Talk by Chad Moore DVM","duration":97,"entryNumber":5}]},{"title":"Multi-lateral holistic data-warehouse","sectionNumber":4,"entries":[{"name":"Talk by Scot Doyle","duration":277,"entryNumber":1},{"name":"Talk by Ally Parisian","duration":112,"entryNumber":2},{"name":"Talk by Miss Marco Feeney","duration":60,"entryNumber":3}]},{"title":"Integrated mobile moratorium","sectionNumber":5,"entries":[{"name":"Talk by Tierra Gislason","duration":30,"entryNumber":1}]},{"title":"Assimilated national parallelism","sectionNumber":6,"entries":[{"name":"Talk by Miss Hobart Hauck","duration":75,"entryNumber":1},{"name":"Talk by Lottie Thompson","duration":40,"entryNumber":2},{"name":"Talk by Roxane Parker","duration":70,"entryNumber":3},{"name":"Talk by Aleen Feeney Sr.","duration":119,"entryNumber":4},{"name":"Talk by Sherman Hand","duration":71,"entryNumber":5},{"name":"Talk by Gayle Hartmann","duration":254,"entryNumber":6},{"name":"Talk by Dr. Chase Cronin","duration":289,"entryNumber":7},{"name":"Talk by Mrs. Maximillia Harris","duration":234,"entryNumber":8}]},{"title":"Persevering human-resource infrastructure","sectionNumber":7,"entries":[{"name":"Talk by Dr. Roselyn Murphy","duration":202,"entryNumber":1},{"name":"Talk by Monroe Auer","duration":346,"entryNumber":2},{"name":"Talk by Tanner Stanton","duration":62,"entryNumber":3},{"name":"Talk by Otto Bergstrom","duration":112,"entryNumber":4},{"name":"Talk by Cyril Beer","duration":110,"entryNumber":5},{"name":"Talk by Mrs. Anastacio Jast","duration":211,"entryNumber":6},{"name":"Talk by Luciano Stoltenberg IV","duration":300,"entryNumber":7},{"name":"Talk by Eric Bernhard","duration":43,"entryNumber":8}]},{"title":"Sharable interactive moratorium","sectionNumber":8,"entries":[{"name":"Talk by Erika Rodriguez","duration":238,"entryNumber":1}]},{"title":"Reactive radical capability","sectionNumber":9,"entries":[{"name":"Talk by Marilyne Schulist","duration":191,"entryNumber":1},{"name":"Talk by Heidi Batz","duration":136,"entryNumber":2},{"name":"Talk by Gudrun Ryan","duration":136,"entryNumber":3},{"name":"Talk by Jack Kuhn Jr.","duration":52,"entryNumber":4}]},{"title":"Cross-group actuating interface","sectionNumber":10,"entries":[{"name":"Talk by Chanelle Stanton","duration":162,"entryNumber":1},{"name":"Talk by Miss Donald Hodkiewicz","duration":260,"entryNumber":2},{"name":"Talk by Hulda Hammes","duration":331,"entryNumber":3},{"name":"Talk by Stephany Kuhn","duration":242,"entryNumber":4},{"name":"Talk by Ms. Raegan Berge","duration":309,"entryNumber":5}]},{"title":"Grass-roots fault-tolerant concept","sectionNumber":11,"entries":[{"name":"Talk by Simeon Cremin","duration":161,"entryNumber":1},{"name":"Talk by Coleman Kuhn","duration":213,"entryNumber":2},{"name":"Talk by Hoyt Heaney","duration":298,"entryNumber":3},{"name":"Talk by Otilia Mitchell","duration":89,"entryNumber":4}]},{"title":"Re-contextualized explicit ability","sectionNumber":12,"entries":[{"name":"Talk by Layne Pagac","duration":164,"entryNumber":1},{"name":"Talk by Mossie Hane","duration":236,"entryNumber":2},{"name":"Talk by Adah Wehner","duration":327,"entryNumber":3}]},{"title":"Cross-platform heuristic knowledge user","sectionNumber":13,"entries":[{"name":"Talk by Ethel Walker","duration":91,"entryNumber":1},{"name":"Talk by Ms. Wilber Schmitt","duration":30,"entryNumber":2},{"name":"Talk by Caden Hilpert","duration":226,"entryNumber":3},{"name":"Talk by Mariana Jaskolski","duration":3,"entryNumber":4},{"name":"Talk by Dr. Sam Heaney","duration":267,"entryNumber":5}]},{"title":"Devolved value-added adapter","sectionNumber":14,"entries":[{"name":"Talk by Ms. Ellen Mraz","duration":315,"entryNumber":1},{"name":"Talk by Carlotta Lemke","duration":257,"entryNumber":2},{"name":"Talk by Paolo Yundt","duration":213,"entryNumber":3},{"name":"Talk by Joe Moore Sr.","duration":264,"entryNumber":4},{"name":"Talk by Dahlia Fay","duration":76,"entryNumber":5},{"name":"Talk by Doug Dibbert","duration":259,"entryNumber":6},{"name":"Talk by Laisha Brown","duration":80,"entryNumber":7},{"name":"Talk by Karelle Maggio","duration":197,"entryNumber":8}]},{"title":"Persistent homogeneous Graphical User Interface","sectionNumber":15,"entries":[{"name":"Talk by Florence Kshlerin V","duration":317,"entryNumber":1},{"name":"Talk by Filomena Hilll","duration":198,"entryNumber":2},{"name":"Talk by Clark Considine","duration":247,"entryNumber":3}]},{"title":"Fully-configurable systematic access","sectionNumber":16,"entries":[{"name":"Talk by Angie Upton","duration":183,"entryNumber":1},{"name":"Talk by Alek Breitenberg","duration":297,"entryNumber":2},{"name":"Talk by Alexys Lynch","duration":165,"entryNumber":3},{"name":"Talk by Donnie Little","duration":15,"entryNumber":4},{"name":"Talk by Derrick Ondricka II","duration":256,"entryNumber":5},{"name":"Talk by Cale Reichel","duration":107,"entryNumber":6},{"name":"Talk by Turner Paucek","duration":139,"entryNumber":7},{"name":"Talk by Raul Hagenes","duration":228,"entryNumber":8}]},{"title":"Enterprise-wide client-driven throughput","sectionNumber":17,"entries":[{"name":"Talk by Michelle Klein","duration":177,"entryNumber":1},{"name":"Talk by Mohammed Bayer","duration":84,"entryNumber":2},{"name":"Talk by Lamont Reilly V","duration":12,"entryNumber":3},{"name":"Talk by Trever Senger","duration":286,"entryNumber":4},{"name":"Talk by Isai Weber","duration":86,"entryNumber":5}]},{"title":"Advanced next generation structure","sectionNumber":18,"entries":[{"name":"Talk by Clovis Marquardt","duration":209,"entryNumber":1},{"name":"Talk by Dr. Aracely Cruickshank","duration":206,"entryNumber":2},{"name":"Talk by Shayne Brekke","duration":193,"entryNumber":3},{"name":"Talk by Gregory Bogan","duration":191,"entryNumber":4},{"name":"Talk by Keeley Walter","duration":14,"entryNumber":5}]},{"title":"Inverse disintermediate policy","sectionNumber":19,"entries":[{"name":"Talk by Laron Hudson","duration":322,"entryNumber":1},{"name":"Talk by Hilton Beier","duration":40,"entryNumber":2},{"name":"Talk by Breana Little","duration":48,"entryNumber":3},{"name":"Talk by Cynthia Moen","duration":348,"entryNumber":4},{"name":"Talk by Ola Raynor","duration":99,"entryNumber":5},{"name":"Talk by Carmen Ortiz","duration":221,"entryNumber":6}]},{"title":"Optimized hybrid emulation","sectionNumber":20,"entries":[{"name":"Talk by Kellen Becker II","duration":31,"entryNumber":1},{"name":"Talk by Julien Wuckert","duration":295,"entryNumber":2},{"name":"Talk by Geovany Mohr","duration":141,"entryNumber":3},{"name":"Talk by Geoffrey McKenzie I","duration":174,"entryNumber":4},{"name":"Talk by Rhiannon White","duration":268,"entryNumber":5}]},{"title":"Ergonomic eco-centric attitude","sectionNumber":21,"entries":[{"name":"Talk by Dannie Sauer","duration":230,"entryNumber":1},{"name":"Talk by Katherine Kuvalis","duration":240,"entryNumber":2},{"name":"Talk by Rocky Runolfsdottir","duration":283,"entryNumber":3},{"name":"Talk by Jonathan Cummings","duration":225,"entryNumber":4},{"name":"Talk by Jaden Schmeler","duration":327,"entryNumber":5},{"name":"Talk by Sidney Koelpin","duration":135,"entryNumber":6}]},{"title":"Customizable didactic paradigm","sectionNumber":22,"entries":[{"name":"Talk by Gage Monahan","duration":51,"entryNumber":1},{"name":"Talk by Maximillian Kuhn","duration":80,"entryNumber":2}]},{"title":"Proactive zero administration flexibility","sectionNumber":23,"entries":[{"name":"Talk by Maynard Howe","duration":202,"entryNumber":1},{"name":"Talk by Mrs. Karine Fisher","duration":133,"entryNumber":2},{"name":"Talk by Ron Russel","duration":101,"entryNumber":3},{"name":"Talk by Opal Gibson","duration":159,"entryNumber":4},{"name":"Talk by Zachariah Smith","duration":248,"entryNumber":5},{"name":"Talk by Herta Kris","duration":108,"entryNumber":6},{"name":"Talk by Alford Auer","duration":91,"entryNumber":7},{"name":"Talk by Marcellus Bogisich","duration":118,"entryNumber":8}]},{"title":"Persevering bi-directional structure","sectionNumber":24,"entries":[{"name":"Talk by Mazie Rice","duration":7,"entryNumber":1},{"name":"Talk by Norris Towne","duration":20,"entryNumber":2},{"name":"Talk by Juana Ferry","duration":161,"entryNumber":3},{"name":"Talk by Meaghan Morar","duration":129,"entryNumber":4}]},{"title":"Realigned asymmetric open architecture","sectionNumber":25,"entries":[{"name":"Talk by Amy Wuckert","duration":128,"entryNumber":1},{"name":"Talk by Agnes Runolfsson","duration":77,"entryNumber":2},{"name":"Talk by Demetris Johnson","duration":66,"entryNumber":3},{"name":"Talk by Claudie McKenzie I","duration":292,"entryNumber":4},{"name":"Talk by Jace Bins DDS","duration":135,"entryNumber":5}]},{"title":"Seamless 24 hour intranet","sectionNumber":26,"entries":[{"name":"Talk by Jacynthe Konopelski","duration":107,"entryNumber":1},{"name":"Talk by Wayne Orn","duration":357,"entryNumber":2},{"name":"Talk by Torrey O'Conner","duration":130,"entryNumber":3}]},{"title":"Advanced non-volatile firmware","sectionNumber":27,"entries":[{"name":"Talk by Miss Hudson Gleason","duration":290,"entryNumber":1},{"name":"Talk by Jairo Beatty","duration":334,"entryNumber":2},{"name":"Talk by Rafael Daniel","duration":338,"entryNumber":3},{"name":"Talk by Evans Braun","duration":343,"entryNumber":4},{"name":"Talk by Casimer Lindgren","duration":311,"entryNumber":5},{"name":"Talk by Ulises Emard","duration":98,"entryNumber":6},{"name":"Talk by Idell Auer","duration":89,"entryNumber":7}]},{"title":"Total client-server frame","sectionNumber":28,"entries":[{"name":"Talk by Unique Kertzmann","duration":179,"entryNumber":1},{"name":"Talk by Otis Frami","duration":2,"entryNumber":2},{"name":"Talk by Tomasa Kassulke","duration":349,"entryNumber":3},{"name":"Talk by Nellie Dach Sr.","duration":273,"entryNumber":4}]},{"title":"Diverse next generation forecast","sectionNumber":29,"entries":[{"name":"Talk by Zoie Gleichner","duration":3,"entryNumber":1},{"name":"Talk by Rebekah Stehr","duration":1,"entryNumber":2},{"name":"Talk by Domenick King","duration":29,"entryNumber":3},{"name":"Talk by Jayson Emard","duration":105,"entryNumber":4},{"name":"Talk by Darrell Daugherty","duration":294,"entryNumber":5},{"name":"Talk by Dock Oberbrunner","duration":309,"entryNumber":6}]},{"title":"Organized eco-centric approach","sectionNumber":30,"entries":[{"name":"Talk by Darby Weissnat","duration":264,"entryNumber":1},{"name":"Talk by Eulalia Johns","duration":290,"entryNumber":2},{"name":"Talk by Delbert Sauer","duration":50,"entryNumber":3}]},{"title":"Switchable leading edge capacity","sectionNumber":31,"entries":[{"name":"Talk by Madisen Cremin","duration":27,"entryNumber":1},{"name":"Talk by Miss Jerad Gerhold","duration":11,"entryNumber":2},{"name":"Talk by Garnett Runte","duration":204,"entryNumber":3},{"name":"Talk by Dino Thompson","duration":190,"entryNumber":4},{"name":"Talk by Ora Streich","duration":46,"entryNumber":5},{"name":"Talk by Kory Konopelski","duration":164,"entryNumber":6},{"name":"Talk by Kameron Skiles","duration":256,"entryNumber":7}]},{"title":"Public-key systematic function","sectionNumber":32,"entries":[{"name":"Talk by Murl Orn","duration":308,"entryNumber":1},{"name":"Talk by George Labadie","duration":264,"entryNumber":2},{"name":"Talk by Cecile Kuhn","duration":252,"entryNumber":3},{"name":"Talk by Edgardo Champlin","duration":320,"entryNumber":4},{"name":"Talk by Jacey Ebert","duration":92,"entryNumber":5},{"name":"Talk by Minerva Dach","duration":156,"entryNumber":6}]},{"title":"Visionary reciprocal methodology","sectionNumber":33,"entries":[{"name":"Talk by Shaniya Block","duration":261,"entryNumber":1},{"name":"Talk by Hester Sporer","duration":350,"entryNumber":2},{"name":"Talk by Claudia Miller","duration":315,"entryNumber":3},{"name":"Talk by Margarett Ferry","duration":81,"entryNumber":4},{"name":"Talk by Ms. Alexis Hessel","duration":75,"entryNumber":5},{"name":"Talk by Ramiro Legros DVM","duration":113,"entryNumber":6}]},{"title":"Networked bottom-line installation","sectionNumber":34,"entries":[{"name":"Talk by Jarret Wolf","duration":106,"entryNumber":1}]},{"title":"Networked even-keeled project","sectionNumber":35,"entries":[{"name":"Talk by Idella Herzog","duration":193,"entryNumber":1},{"name":"Talk by Isabelle Raynor","duration":302,"entryNumber":2},{"name":"Talk by Elody Heller","duration":265,"entryNumber":3},{"name":"Talk by Cristal Ferry","duration":108,"entryNumber":4},{"name":"Talk by Luella Emard","duration":113,"entryNumber":5}]},{"title":"Up-sized logistical adapter","sectionNumber":36,"entries":[{"name":"Talk by Victoria Howe","duration":39,"entryNumber":1}]},{"title":"Implemented uniform firmware","sectionNumber":37,"entries":[{"name":"Talk by Maria Veum","duration":274,"entryNumber":1},{"name":"Talk by Odell Rippin","duration":291,"entryNumber":2}]},{"title":"Virtual transitional support","sectionNumber":38,"entries":[{"name":"Talk by Adelia Brown","duration":322,"entryNumber":1}]},{"title":"Fully-configurable zero administration process improvement","sectionNumber":39,"entries":[{"name":"Talk by Ms. Kiana Aufderhar","duration":345,"entryNumber":1},{"name":"Talk by Muhammad Reinger","duration":55,"entryNumber":2},{"name":"Talk by Eusebio Runte","duration":95,"entryNumber":3},{"name":"Talk by Ocie Crist","duration":266,"entryNumber":4},{"name":"Talk by Beau Lind","duration":84,"entryNumber":5},{"name":"Talk by Yasmine O'Hara II","duration":349,"entryNumber":6}]},{"title":"Upgradable radical analyzer","sectionNumber":40,"entries":[{"name":"Talk by Name Halvorson","duration":65,"entryNumber":1},{"name":"Talk by Ines Sporer","duration":256,"entryNumber":2},{"name":"Talk by Burley Miller","duration":45,"entryNumber":3},{"name":"Talk by Mrs. Andreanne Gerhold","duration":37,"entryNumber":4}]},{"title":"Seamless even-keeled alliance","sectionNumber":41,"entries":[{"name":"Talk by Martin Treutel","duration":121,"entryNumber":1},{"name":"Talk by Dennis Schultz","duration":327,"entryNumber":2},{"name":"Talk by Kyler Wintheiser","duration":325,"entryNumber":3},{"name":"Talk by Devon Farrell","duration":178,"entryNumber":4},{"name":"Talk by Rocky Beatty","duration":265,"entryNumber":5}]},{"title":"User-centric zero tolerance emulation","sectionNumber":42,"entries":[{"name":"Talk by Pietro Kutch I","duration":16,"entryNumber":1},{"name":"Talk by Obie Daugherty","duration":6,"entryNumber":2},{"name":"Talk by Jany Leffler","duration":97,"entryNumber":3},{"name":"Talk by Daren Kiehn","duration":226,"entryNumber":4}]},{"title":"Expanded needs-based secured line","sectionNumber":43,"entries":[{"name":"Talk by Katelyn Mante","duration":55,"entryNumber":1},{"name":"Talk by Tianna Beahan MD","duration":34,"entryNumber":2},{"name":"Talk by Cydney Welch","duration":103,"entryNumber":3},{"name":"Talk by Laury Anderson","duration":294,"entryNumber":4},{"name":"Talk by Freddy Cassin","duration":317,"entryNumber":5},{"name":"Talk by Miss Deshaun Weimann","duration":128,"entryNumber":6},{"name":"Talk by Clarabelle Wiza","duration":299,"entryNumber":7}]},{"title":"Optional fresh-thinking installation","sectionNumber":44,"entries":[{"name":"Talk by Baron Koss","duration":342,"entryNumber":1}]},{"title":"Organized directional alliance","sectionNumber":45,"entries":[{"name":"Talk by Chandler Breitenberg","duration":109,"entryNumber":1},{"name":"Talk by Jaylin Toy","duration":194,"entryNumber":2},{"name":"Talk by Louisa Lindgren","duration":148,"entryNumber":3},{"name":"Talk by Pete Hahn","duration":150,"entryNumber":4},{"name":"Talk by Anabel Sporer","duration":39,"entryNumber":5}]},{"title":"Re-contextualized uniform help-desk","sectionNumber":46,"entries":[{"name":"Talk by Ray Boehm","duration":108,"entryNumber":1},{"name":"Talk by Callie Grant","duration":108,"entryNumber":2},{"name":"Talk by Ollie Dicki","duration":152,"entryNumber":3},{"name":"Talk by Leila Zulauf","duration":323,"entryNumber":4},{"name":"Talk by Haleigh Mante","duration":29,"entryNumber":5}]},{"title":"Visionary impactful hub","sectionNumber":47,"entries":[{"name":"Talk by Mr. Elijah Howell","duration":12,"entryNumber":1},{"name":"Talk by Laverna Kub","duration":314,"entryNumber":2},{"name":"Talk by Keon Baumbach III","duration":162,"entryNumber":3}]},{"title":"Seamless fresh-thinking intranet","sectionNumber":48,"entries":[{"name":"Talk by Franco Bauch","duration":178,"entryNumber":1},{"name":"Talk by Jacey Kovacek","duration":84,"entryNumber":2},{"name":"Talk by Adelle Olson","duration":335,"entryNumber":3},{"name":"Talk by Ashtyn Johns","duration":326,"entryNumber":4},{"name":"Talk by Felix Gibson","duration":345,"entryNumber":5},{"name":"Talk by Merritt Lubowitz MD","duration":49,"entryNumber":6},{"name":"Talk by Mario Crooks","duration":339,"entryNumber":7},{"name":"Talk by Angelo Frami","duration":4,"entryNumber":8}]},{"title":"Configurable reciprocal strategy","sectionNumber":49,"entries":[{"name":"Talk by Lottie Smitham","duration":260,"entryNumber":1},{"name":"Talk by Johnson Carroll","duration":219,"entryNumber":2},{"name":"Talk by Triston Goyette","duration":359,"entryNumber":3},{"name":"Talk by Lucile Sanford","duration":211,"entryNumber":4},{"name":"Talk by Lane O'Reilly","duration":15,"entryNumber":5}]},{"title":"Object-based holistic moratorium","sectionNumber":50,"entries":[{"name":"Talk by Heather Leannon","duration":204,"entryNumber":1},{"name":"Talk by Eula Ernser","duration":193,"entryNumber":2},{"name":"Talk by Ms. Alf Homenick","duration":231,"entryNumber":3},{"name":"Talk by Tyler Welch","duration":19,"entryNumber":4},{"name":"Talk by Eulah VonRueden","duration":347,"entryNumber":5},{"name":"Talk by Carole Collins","duration":147,"entryNumber":6}]},{"title":"Focused executive flexibility","sectionNumber":51,"entries":[{"name":"Talk by Antone Shields","duration":59,"entryNumber":1},{"name":"Talk by Ayla Hahn","duration":299,"entryNumber":2},{"name":"Talk by Cristal Bayer","duration":25,"entryNumber":3},{"name":"Talk by Annetta Cummings","duration":326,"entryNumber":4},{"name":"Talk by Ms. Ophelia Feeney","duration":125,"entryNumber":5},{"name":"Talk by Madisen Bergnaum","duration":27,"entryNumber":6},{"name":"Talk by Carissa Pollich","duration":76,"entryNumber":7}]},{"title":"Fundamental stable Graphic Interface","sectionNumber":52,"entries":[{"name":"Talk by Murl Champlin V","duration":291,"entryNumber":1},{"name":"Talk by Mr. Ramiro Wehner","duration":3,"entryNumber":2},{"name":"Talk by Bryana Haag","duration":170,"entryNumber":3},{"name":"Talk by Johanna Nikolaus","duration":107,"entryNumber":4}]},{"title":"Reverse-engineered multimedia hierarchy","sectionNumber":53,"entries":[{"name":"Talk by Ansel Frami","duration":231,"entryNumber":1}]},{"title":"Open-architected hybrid projection","sectionNumber":54,"entries":[{"name":"Talk by Everardo Heidenreich","duration":2,"entryNumber":1},{"name":"Talk by Susanna Halvorson","duration":359,"entryNumber":2},{"name":"Talk by Kieran Sipes","duration":122,"entryNumber":3},{"name":"Talk by Zackary Lueilwitz","duration":74,"entryNumber":4}]},{"title":"Cross-platform scalable customer loyalty","sectionNumber":55,"entries":[{"name":"Talk by Tyra Kessler","duration":123,"entryNumber":1}]},{"title":"Centralized neutral throughput","sectionNumber":56,"entries":[{"name":"Talk by Reagan Heathcote","duration":344,"entryNumber":1},{"name":"Talk by Marianna Sporer","duration":145,"entryNumber":2},{"name":"Talk by Sydni Upton","duration":326,"entryNumber":3},{"name":"Talk by Janae Brakus","duration":358,"entryNumber":4},{"name":"Talk by Yazmin Hermann","duration":78,"entryNumber":5},{"name":"Talk by Stephan Weber","duration":209,"entryNumber":6},{"name":"Talk by Maia Hickle","duration":257,"entryNumber":7}]},{"title":"Self-enabling analyzing functionalities","sectionNumber":57,"entries":[{"name":"Talk by Merlin Aufderhar DDS","duration":246,"entryNumber":1},{"name":"Talk by Miss Althea Thiel","duration":41,"entryNumber":2},{"name":"Talk by Ms. Beatrice McKenzie","duration":336,"entryNumber":3},{"name":"Talk by Burdette Will","duration":245,"entryNumber":4}]},{"title":"Quality-focused global parallelism","sectionNumber":58,"entries":[{"name":"Talk by Jordon Weber","duration":49,"entryNumber":1},{"name":"Talk by Miss Mitchel Keeling","duration":210,"entryNumber":2},{"name":"Talk by Antonette Rohan","duration":155,"entryNumber":3},{"name":"Talk by Micheal Schultz","duration":42,"entryNumber":4},{"name":"Talk by Christop Raynor","duration":288,"entryNumber":5},{"name":"Talk by Nils Aufderhar","duration":56,"entryNumber":6},{"name":"Talk by Maryjane Watsica","duration":345,"entryNumber":7},{"name":"Talk by Josiah Schultz","duration":37,"entryNumber":8}]},{"title":"Progressive context-sensitive hub","sectionNumber":59,"entries":[{"name":"Talk by Justine Gaylord","duration":295,"entryNumber":1},{"name":"Talk by Enola Bradtke","duration":260,"entryNumber":2},{"name":"Talk by Toney Murphy","duration":145,"entryNumber":3},{"name":"Talk by Leopoldo Franecki","duration":86,"entryNumber":4},{"name":"Talk by Helena Bruen","duration":158,"entryNumber":5},{"name":"Talk by Laisha Hirthe","duration":240,"entryNumber":6}]},{"title":"Robust needs-based strategy","sectionNumber":60,"entries":[{"name":"Talk by Lucio Frami","duration":324,"entryNumber":1},{"name":"Talk by Madison Goldner V","duration":72,"entryNumber":2}]},{"title":"Versatile attitude-oriented migration","sectionNumber":61,"entries":[{"name":"Talk by Rod Schneider","duration":317,"entryNumber":1},{"name":"Talk by Hermina Collier","duration":56,"entryNumber":2},{"name":"Talk by Giovanny Littel","duration":156,"entryNumber":3},{"name":"Talk by Hailee Harvey","duration":196,"entryNumber":4},{"name":"Talk by Charlene Macejkovic","duration":136,"entryNumber":5}]},{"title":"Up-sized clear-thinking time-frame","sectionNumber":62,"entries":[{"name":"Talk by Novella Auer DDS","duration":131,"entryNumber":1},{"name":"Talk by Nash Bauch","duration":98,"entryNumber":2},{"name":"Talk by Issac Veum","duration":115,"entryNumber":3},{"name":"Talk by Lysanne Cormier DVM","duration":353,"entryNumber":4},{"name":"Talk by Ramiro Crona","duration":51,"entryNumber":5},{"name":"Talk by Miss Cooper O'Hara","duration":6,"entryNumber":6},{"name":"Talk by Veronica Hammes","duration":126,"entryNumber":7}]},{"title":"Realigned analyzing matrices","sectionNumber":63,"entries":[{"name":"Talk by Dr. Haylee Barton","duration":270,"entryNumber":1},{"name":"Talk by Ernie Bayer","duration":243,"entryNumber":2},{"name":"Talk by Stephen Kerluke","duration":84,"entryNumber":3},{"name":"Talk by Liliana VonRueden","duration":347,"entryNumber":4},{"name":"Talk by Gregory Heaney","duration":60,"entryNumber":5},{"name":"Talk by Mr. Florida Buckridge","duration":141,"entryNumber":6},{"name":"Talk by Ansel Schoen","duration":344,"entryNumber":7},{"name":"Talk by Newton Stokes","duration":111,"entryNumber":8}]},{"title":"Down-sized logistical success","sectionNumber":64,"entries":[{"name":"Talk by August Bayer","duration":264,"entryNumber":1}]},{"title":"Self-enabling dynamic implementation","sectionNumber":65,"entries":[{"name":"Talk by Coby McLaughlin","duration":132,"entryNumber":1},{"name":"Talk by Ryann Mante","duration":155,"entryNumber":2},{"name":"Talk by Orland Feest","duration":144,"entryNumber":3},{"name":"Talk by Jonathon Jakubowski","duration":336,"entryNumber":4}]},{"title":"Horizontal client-server monitoring","sectionNumber":66,"entries":[{"name":"Talk by Jadon Marquardt","duration":112,"entryNumber":1},{"name":"Talk by Dr. Kaley Bruen","duration":100,"entryNumber":2},{"name":"Talk by Danielle Cronin","duration":108,"entryNumber":3},{"name":"Talk by Victor Abbott","duration":184,"entryNumber":4},{"name":"Talk by Andre Volkman","duration":79,"entryNumber":5},{"name":"Talk by Ms. Emilie Hoeger","duration":79,"entryNumber":6},{"name":"Talk by Ray Jones","duration":25,"entryNumber":7},{"name":"Talk by Gerhard Yost","duration":305,"entryNumber":8}]},{"title":"Balanced interactive implementation","sectionNumber":67,"entries":[{"name":"Talk by Arvid Kuhlman","duration":161,"entryNumber":1},{"name":"Talk by Elmo Bechtelar","duration":236,"entryNumber":2}]},{"title":"Open-source object-oriented support","sectionNumber":68,"entries":[{"name":"Talk by Justice Dach","duration":320,"entryNumber":1},{"name":"Talk by Ms. Maximillian Schinner","duration":72,"entryNumber":2},{"name":"Talk by Jettie Pollich","duration":332,"entryNumber":3}]},{"title":"Progressive real-time workforce","sectionNumber":69,"entries":[{"name":"Talk by Vivienne Zieme","duration":124,"entryNumber":1}]},{"title":"Vision-oriented 24 hour website","sectionNumber":70,"entries":[{"name":"Talk by Shane Robel","duration":29,"entryNumber":1},{"name":"Talk by Leta Kemmer","duration":247,"entryNumber":2},{"name":"Talk by Miss Gina Bergnaum","duration":95,"entryNumber":3},{"name":"Talk by Ellis Kirlin","duration":37,"entryNumber":4}]},{"title":"Fundamental intermediate intranet","sectionNumber":71,"entries":[{"name":"Talk by Mr. Ocie Roob","duration":76,"entryNumber":1},{"name":"Talk by Lowell Hettinger I","duration":139,"entryNumber":2},{"name":"Talk by Stewart MacGyver","duration":210,"entryNumber":3},{"name":"Talk by Francesco Paucek","duration":130,"entryNumber":4}]},{"title":"Down-sized next generation alliance","sectionNumber":72,"entries":[{"name":"Talk by Cody Sanford","duration":247,"entryNumber":1}]},{"title":"Virtual bandwidth-monitored database","sectionNumber":73,"entries":[{"name":"Talk by Abe Jacobi","duration":99,"entryNumber":1},{"name":"Talk by Mabelle Auer","duration":278,"entryNumber":2},{"name":"Talk by Joel Howe","duration":149,"entryNumber":3},{"name":"Talk by Buddy Lind","duration":76,"entryNumber":4},{"name":"Talk by Marcelle Luettgen","duration":180,"entryNumber":5},{"name":"Talk by Oleta Kemmer","duration":348,"entryNumber":6}]},{"title":"Assimilated actuating hierarchy","sectionNumber":74,"entries":[{"name":"Talk by Kirstin Donnelly","duration":258,"entryNumber":1},{"name":"Talk by Candice Pfeffer Sr.","duration":101,"entryNumber":2},{"name":"Talk by Shania Kulas","duration":132,"entryNumber":3},{"name":"Talk by Brooks Grant","duration":129,"entryNumber":4},{"name":"Talk by Monty Mann","duration":279,"entryNumber":5},{"name":"Talk by Lawrence Dare","duration":152,"entryNumber":6},{"name":"Talk by Leonie Dietrich","duration":195,"entryNumber":7},{"name":"Talk by Daphnee Kling","duration":315,"entryNumber":8}]},{"title":"Focused exuding solution","sectionNumber":75,"entries":[{"name":"Talk by Gus Marquardt","duration":279,"entryNumber":1},{"name":"Talk by Queen Gerlach","duration":313,"entryNumber":2}]},{"title":"Distributed dynamic internet solution","sectionNumber":76,"entries":[{"name":"Talk by Kirk Schulist","duration":125,"entryNumber":1}]},{"title":"Multi-lateral heuristic archive","sectionNumber":77,"entries":[{"name":"Talk by Trent Pagac","duration":175,"entryNumber":1},{"name":"Talk by Olin Howe","duration":118,"entryNumber":2},{"name":"Talk by Madilyn Von","duration":311,"entryNumber":3},{"name":"Talk by Rex Ziemann","duration":81,"entryNumber":4},{"name":"Talk by Julia Witting","duration":233,"entryNumber":5},{"name":"Talk by Estevan Collins","duration":195,"entryNumber":6},{"name":"Talk by Dejah Kuhlman","duration":232,"entryNumber":7}]},{"title":"Enhanced multi-tasking moderator","sectionNumber":78,"entries":[{"name":"Talk by Felton Parisian","duration":214,"entryNumber":1},{"name":"Talk by Casimer Ebert","duration":74,"entryNumber":2},{"name":"Talk by Ashly McDermott","duration":299,"entryNumber":3}]},{"title":"Total client-server concept","sectionNumber":79,"entries":[{"name":"Talk by Sandra Cummings","duration":32,"entryNumber":1},{"name":"Talk by Jarrett Hane","duration":117,"entryNumber":2},{"name":"Talk by Mohammed Paucek","duration":113,"entryNumber":3},{"name":"Talk by Dr. Rhea Kulas","duration":6,"entryNumber":4},{"name":"Talk by Adrain Schuppe","duration":113,"entryNumber":5}]},{"title":"Ameliorated scalable analyzer","sectionNumber":80,"entries":[{"name":"Talk by Darian Rutherford","duration":274,"entryNumber":1}]},{"title":"Customizable modular infrastructure","sectionNumber":81,"entries":[{"name":"Talk by Muriel Kemmer","duration":215,"entryNumber":1},{"name":"Talk by Constance Mills","duration":175,"entryNumber":2},{"name":"Talk by Odell Bednar","duration":34,"entryNumber":3},{"name":"Talk by Angel Ziemann","duration":151,"entryNumber":4},{"name":"Talk by Marley Lynch","duration":230,"entryNumber":5},{"name":"Talk by Kyle Larson","duration":119,"entryNumber":6},{"name":"Talk by Bernardo McKenzie","duration":236,"entryNumber":7},{"name":"Talk by Horacio Considine","duration":252,"entryNumber":8}]},{"title":"Decentralized maximized matrices","sectionNumber":82,"entries":[{"name":"Talk by Alberto Lowe","duration":57,"entryNumber":1},{"name":"Talk by Guadalupe Robel MD","duration":110,"entryNumber":2},{"name":"Talk by Granville Borer","duration":123,"entryNumber":3},{"name":"Talk by Mossie O'Hara","duration":25,"entryNumber":4},{"name":"Talk by Erling Stanton","duration":97,"entryNumber":5},{"name":"Talk by Mackenzie Rowe","duration":65,"entryNumber":6},{"name":"Talk by Marianna Senger","duration":187,"entryNumber":7},{"name":"Talk by Malvina Farrell","duration":342,"entryNumber":8}]},{"title":"Robust multi-tasking definition","sectionNumber":83,"entries":[{"name":"Talk by Mrs. Pauline Mertz","duration":60,"entryNumber":1},{"name":"Talk by Adolfo Goyette IV","duration":310,"entryNumber":2},{"name":"Talk by Skylar Schuster","duration":44,"entryNumber":3},{"name":"Talk by Helena Tromp","duration":358,"entryNumber":4},{"name":"Talk by Verla Hessel","duration":4,"entryNumber":5}]},{"title":"Re-contextualized 4th generation solution","sectionNumber":84,"entries":[{"name":"Talk by Mr. Emily Ortiz","duration":327,"entryNumber":1},{"name":"Talk by Clark Welch","duration":52,"entryNumber":2},{"name":"Talk by Lyla Beahan","duration":203,"entryNumber":3},{"name":"Talk by Rudolph Parker","duration":247,"entryNumber":4},{"name":"Talk by Maryam Tremblay","duration":181,"entryNumber":5},{"name":"Talk by Omer Klocko","duration":313,"entryNumber":6},{"name":"Talk by Augustine Thiel","duration":170,"entryNumber":7},{"name":"Talk by Jameson Koch","duration":152,"entryNumber":8}]},{"title":"Cloned secondary synergy","sectionNumber":85,"entries":[{"name":"Talk by Miss Keagan Grimes","duration":102,"entryNumber":1}]},{"title":"Switchable content-based focus group","sectionNumber":86,"entries":[{"name":"Talk by Hillard Hermiston III","duration":156,"entryNumber":1},{"name":"Talk by Florian Feest","duration":27,"entryNumber":2},{"name":"Talk by Anabel Mayert","duration":70,"entryNumber":3},{"name":"Talk by Dr. Morris Leuschke","duration":321,"entryNumber":4}]},{"title":"Pre-emptive 6th generation info-mediaries","sectionNumber":87,"entries":[{"name":"Talk by Tad Krajcik","duration":16,"entryNumber":1},{"name":"Talk by Kyleigh Leannon","duration":105,"entryNumber":2}]},{"title":"De-engineered 5th generation concept","sectionNumber":88,"entries":[{"name":"Talk by Lauryn Bailey","duration":181,"entryNumber":1},{"name":"Talk by Sibyl Douglas","duration":275,"entryNumber":2},{"name":"Talk by Clifton Beatty","duration":82,"entryNumber":3},{"name":"Talk by Gertrude Armstrong Jr.","duration":2,"entryNumber":4},{"name":"Talk by Francisco Flatley","duration":42,"entryNumber":5}]},{"title":"Integrated leading edge complexity","sectionNumber":89,"entries":[{"name":"Talk by Robert Maggio","duration":195,"entryNumber":1},{"name":"Talk by Anita Murazik","duration":269,"entryNumber":2},{"name":"Talk by Clifton Mante","duration":3,"entryNumber":3},{"name":"Talk by Janiya Kshlerin","duration":151,"entryNumber":4},{"name":"Talk by Kip Hudson","duration":27,"entryNumber":5},{"name":"Talk by Mr. Morton Schinner","duration":312,"entryNumber":6}]},{"title":"Mandatory reciprocal budgetary management","sectionNumber":90,"entries":[{"name":"Talk by Aisha Conn","duration":38,"entryNumber":1},{"name":"Talk by Kasandra Wilkinson","duration":119,"entryNumber":2},{"name":"Talk by Akeem Willms","duration":320,"entryNumber":3}]},{"title":"Optional radical function","sectionNumber":91,"entries":[{"name":"Talk by Mrs. Kailey Murray","duration":10,"entryNumber":1},{"name":"Talk by Levi Kub","duration":197,"entryNumber":2},{"name":"Talk by Emie Hermann","duration":75,"entryNumber":3},{"name":"Talk by Chase Reilly","duration":109,"entryNumber":4},{"name":"Talk by Kevin Bernier","duration":273,"entryNumber":5},{"name":"Talk by Dr. Lenora Schamberger","duration":187,"entryNumber":6},{"name":"Talk by Gregorio O'Conner","duration":118,"entryNumber":7}]},{"title":"Triple-buffered foreground intranet","sectionNumber":92,"entries":[{"name":"Talk by Muriel Wolff","duration":285,"entryNumber":1},{"name":"Talk by Blaise Grant","duration":170,"entryNumber":2}]},{"title":"Programmable cohesive intranet","sectionNumber":93,"entries":[{"name":"Talk by Catherine Bernier","duration":352,"entryNumber":1},{"name":"Talk by Laurine Bergnaum","duration":207,"entryNumber":2},{"name":"Talk by Andre Langworth","duration":44,"entryNumber":3},{"name":"Talk by Dr. Vicente Jenkins","duration":193,"entryNumber":4},{"name":"Talk by Mark Morissette","duration":101,"entryNumber":5},{"name":"Talk by Ansel Bartoletti","duration":303,"entryNumber":6},{"name":"Talk by Gus Miller","duration":202,"entryNumber":7}]},{"title":"Reverse-engineered asynchronous help-desk","sectionNumber":94,"entries":[{"name":"Talk by Yoshiko Bayer Jr.","duration":218,"entryNumber":1},{"name":"Talk by Rylee Dare","duration":350,"entryNumber":2},{"name":"Talk by Horacio Abshire","duration":30,"entryNumber":3},{"name":"Talk by Savion Kilback","duration":5,"entryNumber":4},{"name":"Talk by Sam White","duration":278,"entryNumber":5}]},{"title":"Vision-oriented optimal groupware","sectionNumber":95,"entries":[{"name":"Talk by Edison Russel","duration":318,"entryNumber":1},{"name":"Talk by Johathan Kemmer","duration":193,"entryNumber":2}]},{"title":"Optimized dedicated attitude","sectionNumber":96,"entries":[{"name":"Talk by Antonietta Lockman","duration":38,"entryNumber":1},{"name":"Talk by Shannon Wolff","duration":354,"entryNumber":2},{"name":"Talk by Mrs. Elouise Ratke","duration":99,"entryNumber":3},{"name":"Talk by Cordell Fadel","duration":206,"entryNumber":4},{"name":"Talk by Ms. Afton Harber","duration":124,"entryNumber":5},{"name":"Talk by Westley Hickle","duration":45,"entryNumber":6}]},{"title":"Cross-platform dynamic adapter","sectionNumber":97,"entries":[{"name":"Talk by Luisa Runte","duration":41,"entryNumber":1},{"name":"Talk by Alexys Howell","duration":103,"entryNumber":2},{"name":"Talk by Providenci Crona Jr.","duration":72,"entryNumber":3},{"name":"Talk by Rosina Waters IV","duration":293,"entryNumber":4},{"name":"Talk by Izabella Wiza","duration":221,"entryNumber":5}]},{"title":"Pre-emptive eco-centric conglomeration","sectionNumber":98,"entries":[{"name":"Talk by Betty Hauck","duration":212,"entryNumber":1},{"name":"Talk by Salvatore Marvin","duration":171,"entryNumber":2},{"name":"Talk by Kari Morar","duration":355,"entryNumber":3},{"name":"Talk by Olin Streich","duration":211,"entryNumber":4},{"name":"Talk by Marques Yundt","duration":54,"entryNumber":5},{"name":"Talk by Khalid Kris","duration":159,"entryNumber":6}]},{"title":"Public-key dynamic moratorium","sectionNumber":99,"entries":[{"name":"Talk by Emie Okuneva","duration":358,"entryNumber":1},{"name":"Talk by Miguel Effertz","duration":222,"entryNumber":2},{"name":"Talk by Kayleigh West","duration":191,"entryNumber":3},{"name":"Talk by Lane Olson","duration":270,"entryNumber":4},{"name":"Talk by Trycia Streich","duration":69,"entryNumber":5}]},{"title":"Up-sized dynamic alliance","sectionNumber":100,"entries":[{"name":"Talk by Candida Jacobs","duration":256,"entryNumber":1},{"name":"Talk by Shanelle Schamberger DDS","duration":178,"entryNumber":2},{"name":"Talk by Emmanuelle Yost","duration":317,"entryNumber":3},{"name":"Talk by Columbus Osinski","duration":268,"entryNumber":4},{"name":"Talk by Floy Balistreri","duration":216,"entryNumber":5},{"name":"Talk by Delia Schumm","duration":44,"entryNumber":6},{"name":"Talk by Osbaldo Douglas","duration":157,"entryNumber":7}]}] 
  export default courseData;
